requirements.txt
click==8.1.3
pdf2image==1.16.3
Pillow==9.4.0
poppler-utils==0.1.0
